+++ pyTooling/CLIAbstraction/Executable.py
@@ -30,18 +30,13 @@
 # SPDX-License-Identifier: Apache-2.0                                                                                  #
 # ==================================================================================================================== #
 #
-from pathlib              import Path
-from platform             import system
 from subprocess           import Popen				as Subprocess_Popen
 from subprocess           import PIPE					as Subprocess_Pipe
 from subprocess           import STDOUT				as Subprocess_StdOut
-from typing               import Dict, Optional, ClassVar, Type, List
 
 from pyTooling.Decorators import export
-from pyTooling.Exceptions import PlatformNotSupportedException
 
-from . import ExecutableException, DryRunException
-from .Argument import CommandLineArgument, CLIOption, ExecutableArgument, ValuedFlagArgument, NameValuedCommandLineArgument
+from . import ExecutableException, DryRunException, Program
 
 
 @export
@@ -51,98 +46,14 @@ class Environment:
 
 
 @export
-class Program: # (ILogable):
-	"""Represent an executable."""
-	_platform:         str
-	_executableNames:  ClassVar[Dict[str, str]]
-	_executablePath:   Path
-	_environment:      Optional[Dict[str, str]]
-	_dryrun:           bool
-	__cliOptions__:    ClassVar[Dict[Type[CommandLineArgument], Optional[CommandLineArgument]]]
-	__cliParameters__: Dict[Type[CommandLineArgument], Optional[CommandLineArgument]]
-
-	def __init_subclass__(cls, **kwargs):
-		cls.__cliOptions__: Dict[CommandLineArgument, Optional[CommandLineArgument]] = {}
-
-		for option in CLIOption.GetClasses():
-			cls.__cliOptions__[option] = None
-
-	def __init__(self, executablePath: Path = None, binaryDirectoryPath: Path = None, dryRun: bool = False, environment: Environment = None): #, logger : Logger =None):
-		self._platform =    system()
-		self._dryrun =      dryRun
-		self._environment = environment  # if (environment is not None) else Environment()
-
-		if executablePath is not None:
-			if isinstance(executablePath, Path):
-				if not executablePath.exists():
-					raise FileNotFoundError(f"Program '{executablePath}' not found.")
-			else:
-				raise TypeError(f"Parameter 'executablePath' is not of type 'Path'.")
-		elif binaryDirectoryPath is not None:
-			if isinstance(binaryDirectoryPath, Path):
-				try:
-					executablePath = binaryDirectoryPath / self._executableNames[self._platform]
-				except KeyError:
-					raise PlatformNotSupportedException(self._platform)
-
-				if not binaryDirectoryPath.exists():
-					raise FileNotFoundError(f"Binary directory '{binaryDirectoryPath}' not found.")
-				elif not executablePath.exists():
-					raise FileNotFoundError(f"Program '{executablePath}' not found.")
-			else:
-				raise TypeError(f"Parameter 'binaryDirectoryPath' is not of type 'Path'.")
-		else:
-			raise ValueError(f"Neither parameter 'executablePath' nor 'binaryDirectoryPath' was set.")
-
-		self._executablePath = executablePath
-		self.__cliParameters__ = {}
-
-		self.__cliParameters__[self.Executable] = self.Executable(executablePath)
-
-		self._process =  None
-		self._iterator = None
-
-		# if not executablePath.exists():
-		# 	if dryRun:
-		# 		self.LogDryRun(f"File check for '{executablePath}' failed. [SKIPPING]")
-		# 	else:
-		# 		raise ExecutableException(f"Program '{executablePath}' not found.") from FileNotFoundError(str(executablePath))
-
-	def __getitem__(self, key):
-		return self.__cliOptions__[key]
-
-	def __setitem__(self, key, value):
-		if key not in self.__cliOptions__:
-			raise KeyError(f"Option '{key}' is not allowed on executable '{self.__class__.__name__}'")
-		elif key in self.__cliParameters__:
-			raise KeyError(f"Option '{key}' is already set to a value.")
-
-		if issubclass(key, (ValuedFlagArgument, NameValuedCommandLineArgument)):
-			self.__cliParameters__[key] = key(value)
-		else:
-			self.__cliParameters__[key] = key()
-
-	@CLIOption()
-	class Executable(ExecutableArgument, executablePath=None):   # XXX: no argument here
-		def __init__(self, executable: Path):
-			self._executable = executable
-
-	@property
-	def Path(self) -> Path:
-		return self._executablePath
-
-	def ToArgumentList(self) -> List[str]:
-		result: List[str] = []
-		for key, value in self.__cliParameters__.items():
-			result.append(str(value))
-
-		return result
-
-
-@export
 class Executable(Program):  # (ILogable):
 	"""Represent an executable."""
 	_pyIPCMI_BOUNDARY = "====== pyIPCMI BOUNDARY ======"
+
+	def __init__(self, environment: Environment = None):
+
+		self._process =  None
+		self._iterator = None
 
 	def StartProcess(self, parameterList):
 		# start child process
